Remove unnecessary casts in rest of codebase

+++ lpf/include/marlin/lpf/StoreThenForwardBuffer.hpp
@@ -28,12 +28,12 @@ public:
 		if(buf == nullptr) { // Read length
 			if(bytes.size() + size < 8) { // Partial length
 				for(size_t i = 0; i < bytes.size(); i++) {
-					length = (length << 8) | (uint8_t)bytes.data()[i];
+					length = (length << 8) | bytes.data()[i];
 				}
 				size += bytes.size();
 			} else { // Full length
 				for(size_t i = 0; i < 8 - size; i++) {
-					length = (length << 8) | (uint8_t)bytes.data()[i];
+					length = (length << 8) | bytes.data()[i];
 				}
 				bytes.cover(8 - size);
